requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
langchain==0.0.224
pdfplumber==0.7.6
pandas==2.2.3
chromadb==0.4.7
sqlalchemy==2.0.22
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==1.10.12
tqdm==4.66.1
llama-cpp-python==0.1.49
sentence-transformers==2.2.2
